+++ RayTracerB5/Cone.swift
@@ -14,10 +14,9 @@ class Cone: Cylinder {
         var xs = [Intersection]()
         var xsCaps = [Intersection]()
         
-        let a = powf(ray.direction.x, 2) - powf(ray.direction.y, 2) + powf(ray.direction.z, 2)
+        let a = pow(ray.direction.x, 2) - pow(ray.direction.y, 2) + pow(ray.direction.z, 2)
         let b = 2 * ray.origin.x * ray.direction.x - 2 * ray.origin.y * ray.direction.y + 2 * ray.origin.z * ray.direction.z
-        let c = powf(ray.origin.x, 2) - powf(ray.origin.y, 2) +
-            powf(ray.origin.z, 2)
+        let c = pow(ray.origin.x, 2) - pow(ray.origin.y, 2) + pow(ray.origin.z, 2)
 
         if (abs(a) < 0.000001) && (abs(b) < 0.000001) {
             // Ray misses
@@ -34,7 +33,7 @@ class Cone: Cylinder {
             return Intersections(xs)
         }
         
-        var disc = powf(b, 2) - 4 * a * c
+        var disc = pow(b, 2) - 4 * a * c
         if (disc < 0) && abs(disc) > 0.0001 {
             // ray does not intersect the cone, just check Caps
             xsCaps = intersectCaps(ray: ray)
@@ -43,8 +42,8 @@ class Cone: Cylinder {
         }
         
         disc = abs(disc)
-        var t0 = (-b - sqrtf(disc)) / (2 * a)
-        var t1 = (-b + sqrtf(disc)) / (2 * a)
+        var t0 = (-b - sqrt(disc)) / (2 * a)
+        var t1 = (-b + sqrt(disc)) / (2 * a)
         if t0 > t1 { swap(&t0, &t1) }
         
         let y0 = ray.origin.y + t0 * ray.direction.y
@@ -65,7 +64,7 @@ class Cone: Cylinder {
     // a helper function to reduce duplication.
     // checks to see if the intersection at `t` is within a radius
     // of 1 (the radius of your cylinders) from the y axis.
-    func checkCap(ray: Ray, t: Float, r: Float) -> Bool {
+    func checkCap(ray: Ray, t: Double, r: Double) -> Bool {
         let x = ray.origin.x + t * ray.direction.x
         let z = ray.origin.z + t * ray.direction.z
         
@@ -105,7 +104,7 @@ class Cone: Cylinder {
             return Vector(x: 0, y: -1, z: 0)
         }
         
-        var y = sqrtf(powf(p.x, 2) + powf(p.z, 2))
+        var y = sqrt(pow(p.x, 2) + pow(p.z, 2))
         if p.y > 0 { y = -y }
         
         return Vector(x: p.x, y: y, z: p.z)
