+++ RayTracerB5/IntersectionComp.swift
@@ -10,19 +10,19 @@ import Foundation
 
 class IntersectionComp {
    
-    var t: Float
+    var t: Double
     var object: Shape
     var point: Point
     var eyeVector: Vector
     var normal: Vector
     var isInside: Bool
     var reflectv: Vector
-    var n1: Float
-    var n2: Float
+    var n1: Double
+    var n2: Double
     var overPoint: Point
     var underPoint: Point
 
-    init(t: Float, object: Shape) {
+    init(t: Double, object: Shape) {
         self.t = t
         self.object = object
         self.point = Point(x: 0, y: 0, z: 0)
@@ -36,7 +36,7 @@ class IntersectionComp {
         self.underPoint = self.point
     }
     
-    func schlick() -> Float {
+    func schlick() -> Double {
         // find the cosine of the angle between the eye and normal vectors
         var cos = self.eyeVector.dot(self.normal)
         
@@ -47,13 +47,13 @@ class IntersectionComp {
             if sin2_t > 1.0 {
                 return 1.0
             }
-            let cos_t = sqrtf(1.0 - sin2_t)
+            let cos_t = sqrt(1.0 - sin2_t)
             cos = cos_t
         }
         
         let r0 = (self.n1 - self.n2)/(self.n1 + self.n2)
         let r2_0 = r0 * r0
         
-        return r2_0 + (1 - r2_0) * powf((1 - cos), 5)
+        return r2_0 + (1 - r2_0) * pow((1 - cos), 5)
     }
 }
